+++ app/Http/Controllers/OrganizationController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use App\Models\Organization;
+use App\Http\Controllers\AccountController; //check perm
 
 class OrganizationController extends Controller
 {
@@ -15,56 +16,76 @@ class OrganizationController extends Controller
 
     public function show()
     {
-        $orgList = Organization::all();
+        $userperm = AccountController::checkPerm();
+        if($userperm == 1 || $userperm == 2){
+            $orgList = Organization::all();
 
-        #dd($orgList);
+            #dd($orgList);
 
-        return json_encode($orgList);
+            return $orgList->toJson();
+        }
+        else
+            return response(401);
     }
 
     public function store(Request $request)
     {
-        $org = new Organization;
-        
-        $org->name = $request->input('name');
-        $org->contact = $request->input('contact');
-        $org->email = $request->input('email');
-        $org->adress = $request->input('adress');
-        $org->postCode = $request->input('postCode');
-        $org->city = $request->input('city');
-        $org->nip = $request->input('nip');
-        
-        $org->save();
-        
-        return redirect('/Organization/show');
+        $userperm = AccountController::checkPerm();
+        if($userperm == 1 || $userperm == 2){
+            $org = new Organization;
+            
+            $org->name = $request->input('name');
+            $org->contact = $request->input('contact');
+            $org->email = $request->input('email');
+            $org->adress = $request->input('adress');
+            $org->post_code = $request->input('post_code');
+            $org->city = $request->input('city');
+            $org->nip = $request->input('nip');
+            
+            $org->save();
+            
+            return redirect('/organization/show');
+        }
+        else
+            return response(401);
     }
 
     public function update(Request $request)
     {
-        $orgid=$request->input('id');
-        
-        $org = Organization::find($orgid);
+        $userperm = AccountController::checkPerm();
+        if($userperm == 1 || $userperm == 2){
+            $orgid=$request->input('id');
+            
+            $org = Organization::find($orgid);
 
-        $org->name = $request->input('name');
-        $org->contact = $request->input('contact');
-        $org->email = $request->input('email');
-        $org->adress = $request->input('adress');
-        $org->postCode = $request->input('postCode');
-        $org->city = $request->input('city');
-        $org->nip = $request->input('nip');
+            $org->name = $request->input('name');
+            $org->contact = $request->input('contact');
+            $org->email = $request->input('email');
+            $org->adress = $request->input('adress');
+            $org->post_code = $request->input('post_code');
+            $org->city = $request->input('city');
+            $org->nip = $request->input('nip');
 
-        $org->save();
+            $org->save();
 
-        return redirect('/Organization/show');
+            return redirect('/organization/show');
+        }
+        else
+            return response(401);
     }
 
     public function delete(Request $request)
     {
-        $orgid=$request->input('id');
-        
-        $org = Organization::destroy($orgid);
+        $userperm = AccountController::checkPerm();
+        if($userperm == 1 || $userperm == 2){
+            $orgid=$request->input('id');
+            
+            $org = Organization::destroy($orgid);
 
-        return redirect('/Organization/show');
+            return redirect('/organization/show');
+        }
+        else
+            return response(401);
     }
 
 }
