+++ app/Http/Controllers/RealizationController.php
@@ -5,6 +5,11 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use App\Models\Realization;
+use App\Models\Account;
+use App\Models\Employee;
+use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Support\Facades\DB;
+use App\Http\Controllers\AccountController; //check perm
 
 class RealizationController extends Controller
 {
@@ -15,53 +20,71 @@ class RealizationController extends Controller
 
     public function show()
     {
-        $realizationList = Realization::all();
-        return json_encode($realizationList);
+        $userperm = AccountController::checkPerm();
+        if($userperm == 1 || $userperm == 2){
+            $realizationList = Realization::all();
+            return $realizationList->toJson();
+        }
+        else
+            return response(401);
     }
 
     public function store(Request $request)
     {
-        $realization = new Realization;
-        
-        $realization->project = $request->input('project');
-        $realization->employee = $request->input('employee');
-        $realization->startDate = $request->input('startDate');
-        $realization->plannedEndDate = $request->input('plannedEndDate');
-        $realization->supervisor = $request->input('supervisor');
-        
-        $realization->save();
-        
-        return redirect('/Realization/show');
+        $userperm = AccountController::checkPerm();
+        if($userperm == 1 || $userperm == 2){
+            $realization = new Realization;
+            
+            $realization->contract_id = $request->input('contract_id');
+            $realization->start_date = $request->input('start_date');
+            $realization->planned_end_date = $request->input('planned_end_date');
+            $realization->supervisor = $request->input('supervisor');
+            
+            $realization->save();
+            
+            return redirect('/realization/show');
+        }
+        else
+            return response(401);
     }
 
     public function update(Request $request)
     {
-        $realizationid=$request->input('id');
-        
-        $realization = Realization::find($realizationid);
+        $userperm = AccountController::checkPerm();
+        if($userperm == 1 || $userperm == 2){
+            $realizationid=$request->input('id');
+            
+            $realization = Realization::find($realizationid);
 
-        $realization->project = $request->input('project');
-        $realization->employee = $request->input('employee');
-        $realization->startDate = $request->input('startDate');
-        $realization->plannedEndDate = $request->input('plannedEndDate');
-        $realization->supervisor = $request->input('supervisor');
+            $realization->contract_id = $request->input('contract_id');
+            $realization->start_date = $request->input('start_date');
+            $realization->planned_end_date = $request->input('planned_end_date');
+            $realization->supervisor = $request->input('supervisor');
 
-        $realization->save();
+            $realization->save();
 
-        return redirect('/Realization/show');
+            return redirect('/realization/show');
+        }
+        else
+            return response(401);
     }
 
     public function delete(Request $request)
     {
-        $realizationid=$request->input('id');
-        
-        $realization = Realization::destroy($realizationid);
+        $userperm = AccountController::checkPerm();
+        if($userperm == 1 || $userperm == 2){
+            $realizationid=$request->input('id');
+            
+            $realization = Realization::destroy($realizationid);
 
-        return redirect('/Realization/show');
+            return redirect('/realization/show');
+        }
+        else
+            return response(401);
+        
     }
 
-   // public function myRealization($id){
+    
 
-    //}
 
 }
